main 링크경로 변경 , controller 위치 조정
